+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQTF7J359R0BDZ3RZ1BQN8BM"
+        "01JQTM2A44KBMR21ZBH7ES59M1"
     )
     r"""The id of the resource"""