+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQTF7J36D7VM8JNGDXAQ5QCH"
+        "01JQTM2A4514ZFBEHGM2EHDP6T"
     )
     r"""The id of the resource"""