+++ packages/orq-rc/RELEASES.md
@@ -1049,3 +1049,13 @@ Based on:
 - [python v3.3.0-rc.23] packages/orq-rc
 ### Releases
 - [PyPI v3.3.0-rc.23] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.23 - packages/orq-rc
+
+## 2025-04-02 06:48:46
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.527.1 (2.564.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.3.0-rc.24] packages/orq-rc
+### Releases
+- [PyPI v3.3.0-rc.24] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.24 - packages/orq-rc